backend/requirements.txt
# Core Django Framework
Django==5.1.5
djangorestframework==3.15.2

# CORS and Security
django-cors-headers==4.6.0
whitenoise>=6.0.0

# Authentication and Firebase
firebase-admin==6.7.0

# Environment Configuration
python-dotenv==1.0.1

# Image Processing and File Handling
Pillow==11.1.0

# HTTP Requests
requests==2.31.0

# Database
mysqlclient==2.2.4

# Production Server
gunicorn==23.0.0

# Additional utilities (if needed for development/production)
# psycopg2-binary==2.9.9  # PostgreSQL support (uncomment if using PostgreSQL)
# redis==5.0.1            # Redis for caching (uncomment if using Redis)
# celery==5.3.4           # Task queue (uncomment if using Celery)
# django-extensions==3.2.3  # Development utilities (uncomment for development)

# Testing and Development (optional)
# coverage==7.3.2         # Code coverage (uncomment for testing)
# pytest==7.4.3          # Testing framework (uncomment for testing)
# pytest-django==4.7.0   # Django integration for pytest (uncomment for testing)
